+++ src/components/bookertalk/comment/subcomment/SubComment.tsx
@@ -4,10 +4,10 @@ import {
   getSubCommentsInfoHandler,
   insertSubCommentHandler,
   updateSubCommentHandler,
-} from '../../../../api/Supabase1.api';
+} from '../../../../api/Supabase.api';
 import { SubCommentTypes } from '../../../../types/types';
 import { formatCreatedAt } from '../../../../utils/date';
-import * as St from '../Comment.styled';
+import * as St from './SubComment.styled';
 
 type Props = {
   commentId: number | undefined;
