+++ src/components/common/like/Like.styled.ts
