+++ src/components/common/like/PostsLike.tsx
@@ -1,10 +1,10 @@
 import { useCallback, useEffect, useState } from 'react';
-import { getLikeCount, like, supabase } from '../../../api/Supabase1.api';
+import { getLikeCount, like, supabase } from '../../../api/Supabase.api';
 import coloredheart from '../../../assets/common/heavy_black_heart.webp';
 import heartbold from '../../../assets/common/icon_heart.webp';
 import { useAuth } from '../../../contexts/auth.context';
+import * as St from './Like.styled';
 import { LikeProps } from './Like.type';
-import * as St from './Like1.styled';
 
 const PostsLike = ({ postId }: LikeProps) => {
   const [likes, setLikes] = useState<any[]>([]);
