+++ src/components/layout/Layout.tsx
@@ -1,7 +1,7 @@
 import { Outlet } from 'react-router-dom';
 import ChatModal from '../chat/ChatModal';
 import ScrollToTopButton from '../scrolltotopbutton/ScrollToTopButton';
-import * as St from './LayoutStyle';
+import * as St from './Layout.styled';
 import Footer from './footer/Footer';
 import Header from './header/Header';
