+++ src/shared/Router.tsx
@@ -1,11 +1,11 @@
 import { Session } from '@supabase/supabase-js';
 import { useEffect, useState } from 'react';
 import { BrowserRouter, Navigate, Route, Routes } from 'react-router-dom';
-import { getUserSessionHandler } from '../api/Supabase1.api';
+import { getUserSessionHandler } from '../api/Supabase.api';
 import BookDetailPage from '../components/bookintroduction/bookdetailpage/BookDetailPage';
 import AdminChat from '../components/chat/qna/chatadmin/AdminChatRoom';
 import Layout from '../components/layout/Layout';
-import SurveyQuestionnaire from '../components/survey/surveyquestionnaire1/SurveyQuestionnaire';
+import SurveyQuestionnaire from '../components/survey/surveyquestionnaire/SurveyQuestionnaire';
 import SurveyResult from '../components/survey/surveyresult/SurveyResult';
 import AboutBooks from '../pages/AboutBooks';
 import BookerTalk from '../pages/BookerTalk';
